+++ kotlin-sample/src/main/java/io/snabble/sdk/sample/App.kt
@@ -2,6 +2,8 @@ package io.snabble.sdk.sample
 
 import android.app.Application
 import androidx.appcompat.app.AppCompatDelegate
+import com.google.android.material.color.DynamicColors
+import com.google.android.material.color.MaterialColors
 
 class App : Application() {
     companion object {
@@ -11,10 +13,9 @@ class App : Application() {
 
     override fun onCreate() {
         super.onCreate()
-
-        // if you are using a light mode theme, disable night mode resources
-        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-
         instance = this
+
+        // To enable Material You support
+        // DynamicColors.applyToActivitiesIfAvailable(this)
     }
 }