+++ ui/src/main/java/io/snabble/sdk/ui/cart/CheckoutBar.kt
@@ -65,6 +65,8 @@ open class CheckoutBar @JvmOverloads constructor(
     val priceHeight: Int
         get() = priceSum.height + sumContainer.marginTop * 2
 
+    var checkoutPreconditionHandler: CheckoutPreconditionHandler? = null
+
     init {
         orientation = VERTICAL
 
@@ -149,7 +151,9 @@ open class CheckoutBar @JvmOverloads constructor(
     }
 
     protected open fun onPayClick() {
-        pay()
+        if (checkoutPreconditionHandler?.isCheckoutAllowed() != false) {
+            pay()
+        }
     }
 
     private fun update() {
@@ -420,4 +424,8 @@ open class CheckoutBar @JvmOverloads constructor(
             }
         }
     }
+}
+
+fun interface CheckoutPreconditionHandler {
+    fun isCheckoutAllowed(): Boolean
 }